Move shop validation schema to own file
