+++ client/src/components/main/coffee_shop/CoffeeShopForm.js
@@ -2,34 +2,12 @@ import React from 'react';
 import { useMutation } from "@apollo/react-hooks";
 import { Formik, Form, Field, ErrorMessage } from "formik";
 import { useHistory } from "react-router-dom";
-import { object, string, number } from "yup";
 import RenderErrors from "../../util/RenderErrors";
-import './CoffeeShopForm.scss'
+import { ShopSchema } from "./CoffeeShopSchema.js";
+import "./CoffeeShopForm.scss"
 
 import Mutations from "../../../graphql/mutations";
 const { ADD_SHOP } = Mutations;
-
-let ShopSchema = object({
-  name: string()
-    .required("Please enter the name of the cafe"),
-  founded: string()
-    .required("Please enter a year")
-    .matches(/^\d{4}$/ ,"Please enter a valid year"),
-  type: string()
-    .required("Please enter the shop type"),
-  description: string()
-    .required("Please enter description of the cafe"),
-  baristaSatisfaction: number()
-    .required("Barista satisfaction score is required (it's OK to guess)"),
-  address: object({
-    street: string().required("Please enter a street address"),
-    city: string().required("Please enter a city"),
-    state: string().required("Please enter a state"),
-    zip: string()
-      .required("Please enter a ZIP code")
-      .matches(/^[0-9]{5}(?:-[0-9]{4})?$/, "Please enter a valid ZIP code")
-  }).required()
-});
 
 export default (props) => {
   const history = useHistory();
